Updated manager controller with job and products features
