+++ src/main/resources/templates/jobs/completed_jobs.html
@@ -15,8 +15,8 @@
         </li>
         <div>
             <li class="list-item">
-                <a type="button" th:href="@{/parts}" class="btn btn-primary">View Parts</a>
-                <a type="button" th:href="@{/lots}" class="btn btn-primary">View Lots</a>
+                <a type="button" th:href="@{/manager/view_parts}" class="btn btn-primary">View Parts</a>
+                <a type="button" th:href="@{/manager/view_lots}" class="btn btn-primary">View Lots</a>
             </li>
         </div>
     </ul>
