+++ src/main/resources/templates/operation/index.html
@@ -10,7 +10,7 @@
 <header class="container-md header-section" style="padding-top:0.5rem; padding-bottom:0.5rem;">
     <ul class="list" style="list-style:none; display:flex; justify-content:space-between; padding:0; margin: 0;">
         <li class="list-item">
-            <a th:href="@{/manager/product}" class="btn btn-light">Products</a>
+            <a th:href="@{/manager/product/edit/{productId}(productId=${productId})}" class="btn btn-light">Back</a>
         </li>
     </ul>
 </header>
